Bug fixes plus joindtwm command
command: joindtwm added. It sends the discord invite when invoked.
sheets.py bug fixed. The dictionary of columns was wrong, resulting in attendance being put in the wrong column.
Logical error in executeOnEvents fixed. It was checking the current time against a single element in the list of milestones, then waiting if the check failed. Instead, it should now check against the entire list and then wait if the check fails.