Error handler + Attendance testing
The error handler now works.

All Attendance features are tested and complete apart from do_attendance. It's not returning after the 90 minutes is over. I'm working on a fix but progress is slow

I added better logging to the Cog loader.

I made memtils.NameParser synchronous because it didn't really have a reason to be async. It's pretty fast

I renamed DB to Attendance since it also contains the Attendance Cog. I'll move the Cog to its own file if it gets big enough but for now it seems to logically fit in one file