Away commands added, ReactMenu in progress
mark_as_away added to Attendance and mark_away added to AttendanceDBWriter.

All commands that take in a name are now parsed. Regrettably, I wasn't able to implement a decorator for this. This leaves some technical debt

Exist check added to Attendance.add_member. I considered adding this to the DB interface but I decided not to because I wanted to sleep

ReactMenu in progress. It's going to be an interactable Discord.Embed

Fixed a missed f in an f-string in ErrorHandler

Added a designs folder for any design docs