Fixed the DB getting locked
The problem was that the functions executing the queries were async, so the event loop was taking control during the query. This made all of the other query commands wait for the event loop to step back to the query

create_table was made sync to allow this. It didn't make sense to be async anyway because it was mostly being called by async functions

I will not rewrite the DB interface as you're meant to only have one connection with sqlite